chore: refactor to prevent module bundle size from increasing
