+++ src/services/filters.ts
@@ -1,21 +1,6 @@
 import { DetectedLabel } from './fields';
 import { ALL_VARIABLE_VALUE, LEVEL_VARIABLE_VALUE } from './variables';
-import { AdHocFiltersVariable, VariableValueOption } from '@grafana/scenes';
-import { LabelType } from './fieldsTypes';
-import { getLogQLLabelGroups } from './query';
-import { AdHocFilterWithLabels } from './scenes';
-
-export enum FilterOp {
-  Equal = '=',
-  NotEqual = '!=',
-}
-
-export type Filter = {
-  key: string;
-  operator: FilterOp;
-  value: string;
-  type?: LabelType;
-};
+import { VariableValueOption } from '@grafana/scenes';
 
 // We want to show labels with cardinality 1 at the end of the list because they are less useful
 // And then we want to sort by cardinality - from lowest to highest
@@ -56,31 +41,4 @@ export function getFieldOptions(labels: string[]) {
   }));
 
   return [{ label: 'All', value: ALL_VARIABLE_VALUE }, ...labelOptions];
-}
-
-export function joinTagFilters(variable: AdHocFiltersVariable) {
-  const { positiveGroups, negative } = getLogQLLabelGroups(variable.state.filters);
-
-  const filters: AdHocFilterWithLabels[] = [];
-  for (const key in positiveGroups) {
-    const values = positiveGroups[key].map((filter) => filter.value);
-    if (values.length === 1) {
-      filters.push({
-        key,
-        value: positiveGroups[key][0].value,
-        operator: '=',
-      });
-    } else {
-      filters.push({
-        key,
-        value: values.join('|'),
-        operator: '=~',
-      });
-    }
-  }
-
-  negative.forEach((filter) => {
-    filters.push(filter);
-  });
-  return filters;
 }
